chore: reorganize backend tests by responsibility
