+++ backend/tests/presentation/test_openapi.py
