+++ backend/tests/presentation/test_quizzes_api.py
