+++ backend/tests/presentation/test_study_results_api.py
